# platform/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AtivoResposta": {
        "properties": {
          "base64": {
            "title": "Base64",
            "type": "string"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "base64"
        ],
        "title": "AtivoResposta",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API autoritativa da plataforma colaborativa Cursed.",
    "title": "Cursed — Plataforma Colaborativa",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mesas/{mesa_id}/ativos": {
      "get": {
        "description": "``exibicao=true`` entrega a versão reduzida (WEBP) quando ela existe; senão, a original.",
        "operationId": "ler_ativo_mesas__mesa_id__ativos_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "caminho",
            "required": true,
            "schema": {
              "title": "Caminho",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exibicao",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exibicao",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtivoResposta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ler Ativo",
        "tags": [
          "Ativos"
        ]
      }
    },
...
  }
}